pyproject.toml
[project]
name = "todoget"
version = "0.0.1"
readme = "README.md"
description = "A simple python script to track tags (e.g., TODO) within a folder"
authors = [
  { name = "Filippo Gambarota", email = "filippo.gambarota@gmail.com" },
]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
scripts = {todoget = "todoget:todoget"}
requires-python = ">=3.7"

dependencies = [
"click==8.1.7",
"emoji==2.8.0",
"pathspec==0.11.2",
"PyYAML==6.0.1",
]

[project.urls]
"Homepage" = "https://github.com/pypa/sampleproject"
"Bug Tracker" = "https://github.com/pypa/sampleproject/issues"